pyproject.toml
[tool.poetry]
name = "mlops"
version = "0.1.0"
description = ""
authors = ["Daniil Rubin <danyarubin@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^6.1.0"
pre-commit = "^3.6.0"
dvc = "^3.38.1"
scikit-learn = "^1.3.2"
pandas = "^2.1.4"
dvc-s3 = "^3.0.1"
hydra-core = "^1.3.2"
lightgbm = "^4.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
